requirements.txt
llamafactory
transformers
accelerate
tokenizers
torch
datasets
trl
gradio
scipy
einops
sentencepiece
tiktoken
protobuf
uvicorn
fastapi
sse-starlette
matplotlib
fire
omegaconf
packaging
pyyaml
numpy
pydantic
pandas
av
librosa
tyro
